Added window dialogs and refresh list button
Lots of code cleanup, increased functionality for the custom TkWindow and MultiSelectListbox classes.

Issues: Annoying sound from Windows playing whenever a dialog opens, console window flashes briefly on-screen when scanGarbage.exe runs.

+++ src/main.pyw
@@ -3,17 +3,19 @@ import subprocess
 import customTk
 from tkinter import *
 from tkinter import ttk
+from tkinter import messagebox
 from argparse import ArgumentParser
 
 class GarbageRemovalUtilityUI(ttk.Frame):
-    def __init__(self, parent:Widget, padding:int=15):
+    def __init__(self, parent:Widget, scancommand:str, padding:int=15):
+        self.scan_cmd = scancommand
         super().__init__(parent, padding=padding)
         self.grid(column=0, row=0, sticky=(N, S, E, W))
         self.columnconfigure(0, weight=1)
         self.rowconfigure(1, weight=1)
 
         ### MAIN FRAME ###
-        self.label = ttk.Label(self, text="Select the files you want to delete below")
+        self.label = ttk.Label(self, text="Select the file(s) you want to delete below")
         self.label.grid(column=0, row=0, sticky=W)
 
         ### PATHS FRAME ###
@@ -22,58 +24,71 @@ class GarbageRemovalUtilityUI(ttk.Frame):
 
         self.path_listbox = customTk.MultiSelectListbox(
             self.paths_frame,
-            # NOTE TO SELF: Don't hardcode the executable name
-            listvar=self.scanGarbage("scanGarbage.exe")
+            listvar=self.scanGarbage(self.scan_cmd)
         )
         self.path_listbox.grid(column=0, row=0, sticky=(N, S, W, E))
         self.paths_frame.columnconfigure(0, weight=1)
         self.paths_frame.rowconfigure(0, weight=1)
 
-        listbox_scroll = ttk.Scrollbar(
+        self.listbox_scroll = ttk.Scrollbar(
             self.paths_frame,
             orient=VERTICAL,
             command=self.path_listbox.yview
         )
-        self.path_listbox.configure(yscrollcommand=listbox_scroll.set)
-        listbox_scroll.grid(column=1, row=0, sticky=(N, S))
+        self.path_listbox.configure(yscrollcommand=self.listbox_scroll.set)
+        self.listbox_scroll.grid(column=1, row=0, sticky=(N, S))
 
         ### BUTTONS FRAME ###
         self.buttons_frame = ttk.Frame(self)
         self.buttons_frame.grid(column=0, row=2, sticky=(W, E))
         
-        select_button = ttk.Button(
+        self.select_button = ttk.Button(
             self.buttons_frame,
             text="Select all",
             command=self.path_listbox.selectAll
         )
-        select_button.grid(column=0, row=0, sticky=W)
+        self.select_button.grid(column=0, row=0, sticky=W)
 
-        deselect_button = ttk.Button(
+        self.deselect_button = ttk.Button(
             self.buttons_frame,
             text="Deselect all",
             command=self.path_listbox.deselectAll
         )
-        deselect_button.grid(column=1, row=0, sticky=W)
+        self.deselect_button.grid(column=1, row=0, sticky=W)
 
-        delete_button = ttk.Button(
+        self.refresh_button = ttk.Button(
+            self.buttons_frame,
+            text="Refresh list",
+            command=self.refreshListbox
+        )
+        self.refresh_button.grid(column=2, row=0, sticky=W)
+
+        self.delete_button = ttk.Button(
             self.buttons_frame,
             text="Delete selected files",
-            command=self.purgeSelectedFiles
+            command=self.deleteHandler
         )
-        delete_button.grid(column=2, row=0, sticky=E)
-        self.buttons_frame.columnconfigure(2, weight=1)
-    
-    def getPathsFrame(self) -> ttk.Frame:
-        return self.paths_frame
+        self.delete_button.grid(column=3, row=0, sticky=E)
+        self.buttons_frame.columnconfigure(3, weight=1)
 
-    def getButtonsFrame(self) -> ttk.Frame:
-        return self.buttons_frame
+    def refreshListbox(self) -> None:
+        path_list = self.scanGarbage(self.scan_cmd)
+        if path_list == self.path_listbox.getContent():
+            dialog_msg = "No new possible trash detected."
+        else:
+            dialog_msg = "List has been updated with new possible trash."
+            self.path_listbox.setContent(path_list)
+        messagebox.showinfo(
+            title="List Refresh",
+            message=dialog_msg
+        )
     
-    def scanGarbage(self, cmd:str) -> list:
+    def scanGarbage(self, cmd:str=None) -> list:
+        if cmd == None: cmd = self.scan_cmd
         try:
             completed_process = subprocess.run(
                 [cmd],
-                timeout=5,
+                timeout=10,
                 check=True,
                 capture_output=True, 
                 encoding="utf-8"
@@ -84,23 +99,38 @@ class GarbageRemovalUtilityUI(ttk.Frame):
                 return path_set
             return list()
         except TimeoutError:
-            print("Subprocess timed out!")
-        return {"ERROR"}
+            print("ERROR: Subprocess timed out!")
+        return ["ERROR"]
     
-    # NOTE TO SELF: Make accessing of the listbox properties more modular
-    def purgeSelectedFiles(self, *args) -> None:
-        path_query = [e[1] for e in [pair for pair in
-                     enumerate(self.path_listbox.getContent())
-                     if pair[0] in self.path_listbox.curselection()]]
-        for path in path_query:
-            try:
-                os.remove(path)
-                print("REMOVED:", path)
-            except OSError as err:
-                print(f"Error while removing {path}: {err}")
+    def deleteHandler(self, *args) -> None:
+        if len(self.path_listbox.curselection()) == 0:
+            messagebox.showinfo(
+                title="Error",
+                message="Please select one or more files to delete!"
+            )
+            return
+        if messagebox.askyesno(
+            title="Confirmation",
+            message="Are you sure you want to delete these files?"
+        ):
+            path_query = [e[1] for e in [pair for pair in
+                         enumerate(self.path_listbox.getContent())
+                         if pair[0] in self.path_listbox.curselection()]]
+            for path in path_query:
+                try:
+                    os.remove(path)
+                    self.path_listbox.removeContent(path)
+                    print(f"REMOVED: {path}")
+                except OSError as err:
+                    print(f"ERROR REMOVING {path}: {err}")
+            print("QUERY CLEARED")
+            self.path_listbox.deselectAll()
+            messagebox.showinfo(
+                title="Success",
+                message="Files have been deleted."
+            )
 
 if __name__ == "__main__":
-    # NOTE TO SELF: Do something with this argument
     parser = ArgumentParser()
     parser.add_argument(
         "-c", "--command",
@@ -109,7 +139,14 @@ if __name__ == "__main__":
     args = parser.parse_args()
     run_cmd = args.command if args.command != None else "./scanGarbage.exe"
 
-    root = customTk.TkWindow(title="Garbage Removeinator", geometry="500x275")
-    content = GarbageRemovalUtilityUI(parent=root)
-    root.printHierarchy()
+    root = customTk.TkWindow(
+        title="Garbage Removeinator",
+        geometry="500x275",
+        centerscreen=True
+    )
+    content = GarbageRemovalUtilityUI(
+        parent=root,
+        scancommand=run_cmd
+    )
+    # root.printHierarchy()
     root.mainloop()
